CI: wire weekly schedule for asan-codecs leg; gate other jobs
- Add schedule trigger (0 2 * * 0 — weekly Sunday 2am UTC) to workflow on: block
- Gate build-and-test, windows, macos, tests-asan, sonarqube with
  if: github.event_name != 'schedule' so they skip on scheduled runs
- Fix tests-asan-codecs condition: github.event_name == 'schedule' (not .schedule cron)
- tests/image/test_fuzz_codecs.cpp: remove unused #include <array>
- tests/image/test_fuzz_codecs.cpp: cast rng.below(8) to int (line 70)

+++ tests/image/test_fuzz_codecs.cpp
@@ -1,6 +1,5 @@
 #include "test_framework.h"
 
-#include <array>
 #include <cstdint>
 #include <span>
 #include <vector>
@@ -68,7 +67,7 @@ static std::vector<uint8_t> mutate_codec(Prng& rng, const std::vector<uint8_t>& 
     for (uint32_t i = 0; i < num_flips; ++i) {
         if (!v.empty()) {
             const size_t byte_idx = rng.below(static_cast<uint32_t>(v.size()));
-            const int bit_idx = rng.below(8);
+            const int bit_idx = static_cast<int>(rng.below(8));
             flip_bit(v, byte_idx, bit_idx);
         }
     }
